requirements.txt
openai
PyYAML
pandas